+++ example/lib/widgets/multi_feed.dart
@@ -1,10 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:feed/feeds/simple_multi_feed.dart';
 import 'package:tuple/tuple.dart';
+import 'package:sliding_sheet/sliding_sheet.dart';
 
 
 class MultiFeedExample extends StatefulWidget {
-  const MultiFeedExample({ Key? key }) : super(key: key);
+  final SheetController? sheetController;
+
+  const MultiFeedExample({ 
+    Key? key ,
+    this.sheetController
+  }) : super(key: key);
 
   @override
   _MultiFeedExampleState createState() => _MultiFeedExampleState();
@@ -35,6 +41,7 @@ class _MultiFeedExampleState extends State<MultiFeedExample> with TickerProvider
         padding: const EdgeInsets.only(left: 16, right: 16),
         child: SimpleMultiFeed(
           controller: feedController,
+          sheetController: widget.sheetController,
           headerBuilder: (context, i){
             return Column(
               children: [
@@ -79,7 +86,7 @@ class _MultiFeedExampleState extends State<MultiFeedExample> with TickerProvider
                     height: 1,
                     color: Colors.grey.withOpacity(0.2),
                   ),
-                  Container(
+                  SizedBox(
                     height: 75,
                     child: Center(
                       child: Text(list[item % 6], style: const TextStyle(fontWeight: FontWeight.bold, color: Colors.blue))
