--- lib/feeds/page.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-class PageExample extends StatelessWidget {
-  const PageExample({ Key? key }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        child: Text('Hello'),
-      ),
-    );
-  }
-}
